+++ fuge.js
@@ -17,45 +17,14 @@
 
 var program = require('commist')();
 var runner = require('./runner')();
-var gen = require('./generator');
 var shell = require('./shell')();
 var util = require('./util')();
-var minimist = require('minimist');
 var pkg = require('./package.json');
-
-
-
-function argify(args) {
-  return minimist(args, {
-    alias: {
-      c: ['compose-file', 'composefile', 'compose'],
-      i: ['interactive', 'interactivity']
-    },
-    defaults: {
-      i: 1
-    }
-  });
-}
 
 
 
 var showVersion = function() {
   console.log('v' + pkg.version);
-};
-
-
-
-var generateSystem = function(args) {
-  args = argify(args);
-  gen(args.c).generateSystem(args, function() {});
-};
-
-
-
-var generateService = function(args) {
-  args = argify(args);
-  gen(args.c).generateService(args, true, function() {
-  });
 };
 
 
@@ -115,7 +84,6 @@ var previewSystem = function(args) {
 var showHelp = function() {
   console.log('usage: fuge <command> <options>');
   console.log('');
-  console.log('fuge generate <system|service>  generate a system or an additional system service');
   console.log('fuge build                      build a system by executing the RUN commands in each services Dockerfile');
   console.log('fuge pull                       update a system by attempting a git pull against each service');
   console.log('fuge run <compose-file>         run a system');
@@ -127,8 +95,6 @@ var showHelp = function() {
 
 
 
-program.register('generate system', generateSystem);
-program.register('generate service', generateService);
 program.register('build', buildSystem);
 program.register('pull', pullSystem);
 program.register('run', runSystem);